requirements.txt
# Langchain Core
langchain
langchain-core

# OpenAI Integeration
langchain-openai
openai

#Anthoropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integeration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Hugging Face Dependencies
torch
sentence_transformers

# Streamlit for UI
streamlit

#pydantic
pydantic
pydantic[email]